fix: resolve all clippy lints for CI pass
- Refactor walk_dir to use WalkState struct (fixes too_many_arguments, only_used_in_recursion)
- Collapse nested if in pyproject.toml detection (collapsible_if)
- Collapse nested if in input history (collapsible_if)
- Remove redundant u16 cast on scroll_offset (unnecessary_cast)
- Add stray crate dirs to .gitignore

+++ crates/fever-tui/src/screens/chat.rs
@@ -94,7 +94,7 @@ pub fn render(f: &mut Frame, area: Rect, state: &mut AppState) {
     let messages_para = Paragraph::new(lines)
         .block(messages_block)
         .wrap(Wrap { trim: false })
-        .scroll((state.scroll_offset as u16, 0))
+        .scroll((state.scroll_offset, 0))
         .style(Style::default().fg(theme.fg()).bg(theme.bg()));
     f.render_widget(messages_para, chunks[0]);
